+++ lib/Controller/AbstractCheckoutRecurringController.php
@@ -1,11 +1,11 @@
-<?php namespace Vankosoft\PaymentBundle\Controller\PaymentMethod;
+<?php namespace Vankosoft\PaymentBundle\Controller;
 
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpKernel\Exception\HttpException;
 
 use Payum\Core\Request\GetHumanStatus;
 
-abstract class AbstractRecurringPaymentMethodController extends AbstractPaymentMethodController
+abstract class AbstractCheckoutRecurringController extends AbstractCheckoutController
 {
     
     abstract public function createRecurringPaymentAction( $packagePlanId, Request $request );