+++ lib/Controller/Checkout/PaypalExpressCheckoutRecurringController.php
@@ -1,4 +1,4 @@
-<?php namespace Vankosoft\PaymentBundle\Controller\PaymentMethod;
+<?php namespace Vankosoft\PaymentBundle\Controller\Checkout;
 
 use Payum\Bundle\PayumBundle\Controller\PayumController;
 use Payum\Core\Request\Cancel;
@@ -10,6 +10,7 @@ use Payum\Core\Request\GetHumanStatus;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpKernel\Exception\HttpException;
 
+use Vankosoft\PaymentBundle\Controller\AbstractCheckoutController;
 
 use IA\UsersBundle\Entity\UserActivity;
 
@@ -19,7 +20,7 @@ use IA\UsersBundle\Entity\UserActivity;
  * -----------------------------------------------
  * sb-wsp2g401218@personal.example.com / 8o?JWT#6
  */
-class PaypalExpressCheckoutRecurringController extends AbstractPaymentMethodController
+class PaypalExpressCheckoutRecurringController extends AbstractCheckoutController
 {
 
     public function prepareAction( Request $request )