+++ lib/Controller/Checkout/StripeCheckoutController.php
@@ -1,8 +1,10 @@
-<?php  namespace Vankosoft\PaymentBundle\Controller\PaymentMethod;
+<?php  namespace Vankosoft\PaymentBundle\Controller\Checkout;
 
 use Symfony\Component\HttpFoundation\Request;
 
-class StripeCheckoutController extends AbstractPaymentMethodController
+use Vankosoft\PaymentBundle\Controller\AbstractCheckoutController;
+
+class StripeCheckoutController extends AbstractCheckoutController
 {
     /*
      * TEST MAIL: i.atanasov77@gmail.com