+++ lib/Form/Type/GatewayConfigType.php
@@ -22,19 +22,21 @@ class GatewayConfigType extends AbstractType
     }
 
     public function buildForm(FormBuilderInterface $builder, array $options)
-    {   
-        
+    {
         $builder
             ->add('factory', HiddenType::class)
         ;
-        if(isset($options['data']['sandbox'])) {
-            $builder->add('sandbox', HiddenType::class);
+        
+        if( isset( $options['data']['sandbox'] ) ) {
+            $builder->add( 'sandbox', HiddenType::class );
         }
         
-        foreach($options['data'] as $optKey => $optVal) {
-            if(in_array($optKey, array('factory', 'sandbox')))
+        foreach( $options['data'] as $optKey => $optVal ) {
+            if( in_array( $optKey, ['factory', 'sandbox'] ) )
                 continue;
-            $builder->add($optKey, TextType::class, array('attr' => array('size'=>100)));
+            
+            //$value  = in_array( $optKey, $options['values'] );
+            $builder->add( $optKey, TextType::class, [ 'attr' => ['size'=>100] ] );
         }
     }
 }