+++ lib/Model/Interfaces/PaymentInterface.php
@@ -1,8 +1,12 @@
 <?php namespace Vankosoft\PaymentBundle\Model\Interfaces;
 
 use Sylius\Component\Resource\Model\ResourceInterface;
+use Payum\Core\Model\PaymentInterface as BasePaymentInterface;
+use Payum\Core\Model\DirectDebitPaymentInterface;
 
-interface PaymentInterface extends ResourceInterface
+interface PaymentInterface extends ResourceInterface, BasePaymentInterface, DirectDebitPaymentInterface
 {
-    
+    public function getUser();
+    public function getPaidServicePeriod();
+    public function getPaymentMethod();
 }